% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{ph2bayes-package}
\alias{ph2bayes-package}
\alias{ph2bayes}
\title{The 'ph2bayes' package}
\description{
An implementation of Bayesian single-arm phase II design methods
for binary outcome based on posterior probability and predictive
probability:
Thall and Simon (Biometrics 1994),
Lee and Liu (Clinical Trials 2008).
}
\keyword{design}
\keyword{htest}
\keyword{package}
